Refactor and final commit
-Broke out some helper methods to help pojos

-would continue to refactor with time
