+++ lib/handlers.js
@@ -1,5 +1,4 @@
 const $ = require('jquery')
-import COLORS from './data/colors'
 
 
   function showTopColor(topColorObject) {
