+++ lib/pojos.js
@@ -1,23 +1,16 @@
 const $ = require('jquery')
 const handlers = require('./handlers')
 const requests = require('./requests')
+const helpers = require('./helpers')
 import COLORS from './data/colors'
 
 
-Array.prototype.unique = function() {                   //srsly, javascript, por que
-  return this.filter(function (value, index, self) {
-    return self.indexOf(value) === index;
-  });
-}
-
-
-function addSwatches(input) {
+function addSwatches(input) {  // would continue refactoring into helpers
   let colors = input.split(' ')
-  colors = makeLowerCase(colors)
+  colors = helpers.makeLowerCase(colors)
   let colorDict = COLORS
   let colorKeys = Object.keys(colorDict)
   let uniqColors = colors.unique()
-
 
   uniqColors.forEach (function(color) {
     if (colorKeys.includes(color)) {
@@ -31,14 +24,6 @@ function addSwatches(input) {
     }
   })
 
-}
-
-function makeLowerCase (colors) {  //move to helper
-  let sanitizedArray = []
-  colors.forEach( function (color) {
-    sanitizedArray.push(color.toLowerCase())
-  })
-  return sanitizedArray
 }
 
 export {addSwatches}
